# Conflicts:
#	src/core/Piece.java
#	src/ui/Screen.java

+++ src/core/HintManager.java
@@ -226,8 +226,8 @@ public class HintManager {
                 }
             }
             p.setCircleOffsets( newOffsets );
-            p.setLocation( new Point(   J * Ground.EDGE_LENGTH + p.getLevel().getMinX() + Ground.EDGE_LENGTH / 2,
-                                        I * Ground.EDGE_LENGTH + p.getLevel().getMinY() + Ground.EDGE_LENGTH / 2) );
+            p.setLocation( new Point(   (int)(J * Ground.EDGE_LENGTH + p.getLevel().getMinX() + Ground.EDGE_LENGTH / 2),
+                                        (int)(I * Ground.EDGE_LENGTH + p.getLevel().getMinY() + Ground.EDGE_LENGTH / 2)) );
             p.recalculatePoints();
 
         } else {
